Rearranging tasks and adding native run
