+++ tasks/__init__.py
@@ -2,7 +2,6 @@ from invoke import Collection
 
 from . import container
 from . import data
-from . import mpi
 from . import native
 from . import run
 from . import wasm
@@ -10,7 +9,6 @@ from . import wasm
 ns = Collection(
     container,
     data,
-    mpi,
     native,
     run,
     wasm,
